[DF-9566] Warning fix.
Former-commit-id: 9c9d5c46d8cf19100814c1b918462817112d3b64

+++ Sources/Internal/Platform/TemplateWin32/MovieViewControlWin32.cpp
@@ -124,7 +124,7 @@ void MovieViewControl::Update()
     }
     else
     {
-        videoTexture->TexImage(0, textureWidth, textureHeight, videoTextureBuffer, drawData.data.size(), Texture::INVALID_CUBEMAP_FACE);
+        videoTexture->TexImage(0, textureWidth, textureHeight, videoTextureBuffer, static_cast<uint32>(drawData.data.size()), Texture::INVALID_CUBEMAP_FACE);
     }
 }
